+++ R/Evaporative.R
@@ -1,7 +1,7 @@
-#' Construction function for class "Evaporative"
+#' DEPRECATED - Construction function for class "Evaporative"
 #'
-#' @description \code{Evaporative} returns a tranformed object with class "Evaporative" and
-#' units g/day. This class represents the daily emissions presented by
+#'@description \code{Evaporative} returns a tranformed object with class "Evaporative" and
+#' units g. This class represents the daily emissions presented by
 #' Mellios G and Ntziachristos (2016) Gasoline evaporation, Tier 2. Eventually
 #' it will be incorporated the techniques of Tier 3.
 #'
@@ -14,15 +14,15 @@
 #' @rdname Evaporative
 #' @aliases Evaporative print.Evaporative summary.Evaporative
 #' plot.Evaporative
-#' @examples {
-#' ef1 <- ef_evap(ef = "erhotc",v = "PC", cc = "<=1400", dt = "0_15", ca = "no")
-#' ef1
-#' }
 #' @export
+#' @examples \dontrun{
+#' # do not run
+#' # DEPRECATED
+#' }
 Evaporative <- function(x, ...) {
   ev <- x
 if ( is.numeric(ev) ) {
-    ev <- ev*units::as_units("g d-1")
+    ev <- ev*units::as_units("g")
     class(ev) <- c("Evaporative",class(ev))
   }
   return(ev)